Add separate scanHandler
